requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
openai>=1.12.0
python-dotenv==1.0.0
pillow>=10.2.0
requests==2.31.0
qrcode[pil]>=7.4.2

